Rename TracerPlugin to TelemetryPlugin
with the support of metrics the TracerPlugin name is no longer
adequate. Renaming this to TelemetryPlugin.
Also introducing TelemetryProvider interface. While it is only used in
Node.java at the moment to fetch Tracer instance, it is intended to be
used in Plugin::createComponents (to be done in separate commite due to
the broad scope of this method)
This will allow for plugins to get access to both Tracer and Metric interfaces
without the need to add yet another argument to createComponents

Also adding internal subpackage in module/apm so that it is more obvious
which packages are not exported
