+++ server/src/main/java/org/elasticsearch/plugins/TelemetryPlugin.java
@@ -10,7 +10,6 @@ package org.elasticsearch.plugins;
 
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.telemetry.TelemetryProvider;
-import org.elasticsearch.telemetry.tracing.Tracer;
 
 public interface TelemetryPlugin {
     TelemetryProvider getTelemetryProvider(Settings settings);
